rename file with types
